yippe

+++ Practice Proggram Batch 2/quotient.py
@@ -1,5 +1,5 @@
 # Print the quotient of the two numbers without the decimal point
 x = float(input('Enter a number: '))
 y = float(input('Divide with?: '))
-h = x / y
+h = int(x / y)
 print(h)
